git: switch default `push` mode to "simple"
In this mode, when a local branch is pushed to the remote, it will use a
remote branch of the same name if one exists. Otherwise, the operation
will fail, requiring an explicit branch name.

This is now the default in git 2.0+ and is safer than "current", which
would act similarly but also create the new remote branch if one didn't
already exist. I may consider moving to "current" at some point.

I'm moving away from "upstream" because there's never really been a case
when I wanted to use different names for my local and remote branches.