vim: prefer `git ls-files` for ctrlp_user_command
In git repositories, `git ls-files` is much faster than using other
external commands (especially for large repositories). The previous
approach of selectively setting `ctrlp_user_command` to a string
unilaterally used the external command and skipped the git repository
optimization.

Instead, use the expanded form of `ctrlp_user_command` (there are three
forms in total) that accepts a fallback command that is used when we're
*not* in a git repository.