vim: move spelling exceptions to after/syntax
This appears to be the most reliable way to add spelling exceptions to
the @NoSpell cluster. The downside to this approach is that these will
need to be duplicated for each syntax that wants them.

I'm starting with just the gitcommit syntax because it's easy to add
these syntax matches to the top level, but other syntax might require
more specific containment modifiers (e.g. `containedin=`).

If this becomes too much down the road, I'll consider whether there's
something that could be done with the Syntax autocmd that might make
this more general.

... but for now, I'm primarily interested in adding these exceptions to
the gitcommit syntax, and that itch is now scratched.