vim: improve gitcommit "no spell" matches
The previous approach had a couple of issues:

 1. It didn't cover the gitcommitSummary region.
 2. The use of \p in the "multicaps" pattern was too aggressive.

This revised approach defines matches for both the commit body and the
contained commit summary line. I couldn't find a way that let me define
a single definition that would work as both a contained transparent
match as well as a top-level body match.

I've rewritten the "multicaps" pattern to cover all mixed-case cases
(such as camelCase and ALLCAPS). There's also a new pattern that matches
underscores and dots that disables spelling for 'words.like_this'.

Lastly, I prefixed the match names with 'gitcommit' for consistency.