vim: add custom 'compiler' support to projections
Neither of Projectionist's builtin 'make' and 'dispatch' properties can
be used to simply set :compiler, but that would be useful functionality
for e.g. using `exunit` to run Elixir tests.

Fortunately, it's easy to add this support myself by responding to the
ProjectionistActivate autocommand and querying for a custom 'compiler'
value.