vim: add the asyncrun plugin
This plugin provides a simple :AsyncRun command that executes a command
asynchronously (using vim's job control system) and displays its output
in the quickfix window using the current &errorformat value.

I also added basic lightline support that includes the current command
name in the statusbar when an asynchronous job is running.

I don't anticipate using this very often so the plugin is only loaded
on-demand.