vim: roll my own status line highlight functions
This drops the dependency on wincent/pinnacle, which previously supplied
similar functions.

This was largely an exercise in learning how to manipulate highlights
and color groups. The result is a bit more code than was there before,
but with one less external dependency. In its current form, this isn't
any more or less capable than the previous version.