vim: define custom vim-sandwich textobj mappings
I didn't like how the default -query mappings overwrote vim's built-in
sentence-wise text objects, which are useful when writing prose, so I
now map these objects to `iq` and `aq`.

I also considered `iS` and `aS`, or even not defining -query mappings at
all (because I rarely use them), but this is a good compromise for now.

See machakann/vim-sandwich#62