vim: GitFiles now used :FZF outside of git repos
It's convenient for my main mapping (<leader>\) to behave the same both
inside and outside of git repositories. We can't emulate the GitFiles?
(modified files) behavior so that still returns an error outside of a
repository.

Also, this makes my previous <leader>f mapping obsolete so replace it
with a command that opens fzf in the current buffer's directory.