vim: replace ferret with my own grep routines
It's easy to replicate this plugin's functionality in a minimal amount
of code. The key is to execute &grepprg in a sub-shell, an approach
inspired by this gist:

https://gist.github.com/romainl/56f0c28ef953ffc157f36cc495947ab3

This ended up being is as good as an synchronous job for my purposes.

Building on that, I added some title-setting code and light integration
with my status line. I'm using the same mappings as before, but they
reference the new :Grep command instead of the old :Ack command family.