vim: revise buffer switching / deleting maps
A recent vim 8.2 patch revision caused vim-bbye to crash. While that's
since been fixed (by a later vim patch), it caused me to investigate how
I might write some simple mappings that produced similar functionality.

The result is an expression-based `<leader>d` map that is based on the
old `:bp|bd#` trick of switching away from the current buffer and then
deleting the alternate buffer. It avoids the problem of closing a window
that would become empty as a result by conditionally creating a new,
empty buffer if we're about the delete the only remaining listed buffer.

Also, while I'm here, teach the `<tab>` mapping (which switched to the
alternate buffer) to use `:bnext` when when the alternate buffer is no
longer loaded. This "fixes" the problem where this map would reopen a
buffer that was just deleted.