vim: improve GitFiles? output and diff previews
This function "merges" the output of two git commands to present a
unified view of changed files that were already committed on the current
branch plus files that are modified in the working tree. Before this
change, a file that appear in both sets would appear twice, and items
represented committed changes didn't show a diff preview.

We now collapse duplicate filenames entries with the help of `awk`. When
a filename is both committed and modified, we prefer the modified entry.
This changes the sort order from what we had previously, but the result
now has the small benefit of changing the initial selection to a more
recently modified file.

The preview command now explicit diffs each file against `origin`, which
gives us the correct diff preview in all cases I tried. For filenames
that have committed and pending changes, the preview diff shows both
sets of changes (relative to origin), which is a nice improvement.

Lastly, I combined the `cat filename | head -500` command pair into a
single call to `head -500 filename`. That's shorter and (slightly)
faster.