vim: introduce a new notes plugin
This is an extremely simply Notational Velocity-style notes system built
around a single :Notes command which lists the contents of the ~/Notes
directory using fzf.

If the fzf query doesn't match an existing file, a new Markdown file
named after the query string will be created in the ~/Notes directory.