vim: disable g:gutentags_generate_on_new
This was often resulting in a complete regeneration of the tags file.
That appears to happen because I would start a vim session with an
"unknown" file (either an empty buffer or a scratch file), and upon
opening a "known" file, gutentags would consider it as a "new" file and
trigger a "world project" generation.

I'm not sure if this is the intended behavior, but I also haven't been
relying on this "generate on new" functionality, so it's easy to just
disable it and move on.