bash: indicate background jobs in prompt
This is a small experiment that I've been enjoying lately. It adds a `*`
to the prompt when there's at least one job running in the background.
I tried out some other characters as the indicator, but `*` mirrors the
`git status` prompt to indicate "pending work" and is otherwise not too
distracting.

It's common for people to include the full job count (`\j`) in the
prompt, but I don't want/need to see that level of detail. A simple
binary indicator is good enough for me.