vim: revise 'statusline' background colors
With this change, I now build all of my custom 'statusline' highlight
groups off of the StatusLine background color. This results in better
results across a variety of color schemes because it matches the
intention of the schemes' designers. For those cases where I might want
a different background color, I'll customize StatusLine itself.

This code is still more complicated than I'd like. It's doing a lot of
"advanced" work to produce otherwise simple highlights and decorations.
The fairly new hlget() and hlget() functions will help a bit, but I'm
not sure if I want to rely on them just yet.