git: remove ability to receive `files` options
I haven't found a good way to allow `files` to receive both file
arguments (which required $GIT_PREFIX processing) and additional
`ls-files` arguments. The former is more important to me at the moment
so drop support for the latter (and assume `-m` [modified]).

I would ideally like write a brace expansion that only adds $GIT_PREFIX
to those arguments that don't start with -:

    ${@/#[^-]/$GIT_PREFIX}

Unfortunately, that approach consumes the first character of each match,
resulting in `$GIT_PREFIX/oo` instead of `$GIT_PREFIX/foo`. Bash offers
a `&` replacement token containing the full match, letting us do:

    ${@/#[^-]*/$GIT_PREFIX&}

... but that's only available via the `patsub_replacement` shell option
that was introduced in bash 5.2, and I still use some systems with older
versions. I'll hang on to this solution for the future, when this option
is more widely available to me.

https://www.gnu.org/software/bash/manual/html_node/Shell-Parameter-Expansion.html

Lastly, prefer `$@` (array) to `$*` (string) for these operations.