bash: defer homebrew bash completion loading
In e94c50e3, I loaded the bash completion scripts earlier so that
__git_complete would be available for use with the `gv` function.

Unfortunately, this "broke" completion for directories in CDPATH, which
I generally define in `.bashrc.local`, due to the new ordering.

To fix that, go back to loading the bash completion scripts later, but
explicitly load just git-completion.bash before we need it for our local
functions.