git: adopt ~/.config/git for all git configuration
This directory is automatically recognized by git as a configuration
file source. It also means I don't have to explicitly specify the global
ignore file (`core.excludesfile`) anymore.

It will probably take some time for my muscle memory to update to this
new location. I considered leaving just ~/.gitconfig where it was, but I
ultimately decided to go all-in on this approach.