git: remove $GIT_PREFIX behavior from 'mine'
As previously noted, this make it difficult to pass additional 'log'
arguments to this alias (because of the --). I could write a more
complex shell script implementation, but for now just do the simple
thing and require more explicit path ref arguments for when this command
is run from within subdirectories (e.g. `git mine .`).