vim: enable vim-elixir for markdown buffers
Some recent change (that I haven't been to isolate) appears to require
this plugin, which provides the Elixir syntax (syntax/elixir.vim), to be
available in order for `elixir` to appear in the list of fenced markdown
languages (`g:markdown_fenced_languages`).

It was previously being lazily loaded (on demand).