vim: remove 'GitFiles' preview
It's more useful to use the full screen real estate to find and narrow
the filename. Seeing the contents of the files isn't that relevant to
this kind of operation.

Also, move the 'GitFiles?' (modified files) preview above the file list,
which gives both more horizontal space for normally-sized terminal
windows.