vim: always 'set background=dark' in vim
Vim defaults to background=light in most terminal environments. It then
attempts to query the terminal for the background color ('t_RB') and
then adjusts 'background' accordingly _if it hasn't already been set_.
Because this query happens _after_ .vimrc is read, we've already set
'colorscheme', this unfortunately causes a light-to-dark redraw flash if
the color scheme provides both light and dark backgrounds.

To avoid that annoyance, always use a dark background. Neovim doesn't
have the same problem, perhaps due to the order and timing in which it
queries the terminal relative to when 'colorscheme' is set and applied.