vim: a few revisions to the :Grep command
First, pass the arguments expandcmd() so that we get automatic expansion
of characters like '%' (current file). These could already be expanded
on the command line (via <tab>). (I'll keep on eye on whether or not
this interferes with any search patterns; I might need to mix in a call
to fnameescape(), too.)

Second, use <f-args> instead of <q-args> to pass the command arguments.
The "functions args" version is what we want because we're wrapping a
function call. The code was already expecting this, too; with the
previous <q-args> behavior, all of our arguments were being passed as
a:000[0] anyway.

Lastly, restrict completed file names to those within 'path'.