git: restore previous behavior of 'edit'
20ec6d4 added committed branch changes to the set of files opened in the
editor. After using it for a few days, I think I prefer the previous
behavior: only open the set of local (uncommitted) changes. I still have
the ability to open the full set of changes in vim itself.

I did make one revision to the previous behavior: use `xargs -r` to only
open the editor when there are files to edit. I'll see how I like this
in practice because previously `git edit` was also a shortcut to open
the editor in any state.