vim: attempt to simplify some git operations
First, use FugitiveWorkTree() to find the root directory. This has the
advantage of being buffer-aware, for those less frequent cases when I'm
working on files outside of the current working directory.

Second, rather than resolving the merge base commit, use origin/HEAD
plus the "triple dot" notation to identify the commits that our branch
does not have in common with origin/HEAD. This is the simplest approach
that covers the vast majority of cases, but if I ever wanted to handle
cases like branches-of-branches, I'll need to do a bit more work. Git
unfortunately doesn't provide a notation for those topologies.