vim: simplify color schemes
Remove the 'embark' and 'edge' color schemes. I primarily used 'edge'
for its light theme, but the 'retrobox' light theme is pretty good and
ships with vim.

Also, default to the 'quiet' theme when 'termguicolors' is unavailable.
'spring-night' also works well in this mode, but 'quiet' is built-in so
it makes a good fallback.