vim: remove python matchit configuration
Unfortunately, in its current form, matchit is not able to support
Python's `if` syntax. It expects there to be a reliable "end" pattern,
whereas Python allows either `elif` and `else` to be the final keyword
(which zero or more `elif`s in between).

Given that, remove this partial support until its possible to implement
a more complete solution.