vim: add visual mode :Grep mappings
These borrow the "z register because the current visual selection isn't
stored in a standard register. (The clipboard selection is in "+, but
that relies on the system and 'clipboard' setting, so I decided not to
depend on that.)

We also shellescape() the search string because it can possibly contain
spaces and other "special" characters (unlike the word-based normal mode
mappings).