vim: add vim-flog for git branch viewing
This plugin is faster and more richly featured than gv.vim. I haven't
decided which I prefer yet. gv.vim is does *nearly* everything that I
want, which only a few shortcomings that vim-flog addresses:

- vim-flog has a prettier glyph-based display
- vim-flog's navigation commands are a little bit richer
- vim-flog is overall faster by using `git commit-graph write`

I'm tempted to hack on gv.vim and improve it with some of these things,
and I still might do that, but in the mean time, it's nice to just use
something "off the shelf".